Split up TypeDbDatastore into Mixins for better maintenance
