+++ pyproject.toml
@@ -1,30 +1,32 @@
 [project]
 authors = [
-  {name = "Hoog-v", email = "hogeweyv@gmail.com"},
-  {name = "Simon Dingemans", email = "simon.dingemans@gmail.com"}  
+  { name = "Hoog-v", email = "hogeweyv@gmail.com" },
+  { name = "Simon Dingemans", email = "simon.dingemans@gmail.com" },
 ]
 name = "database_builder_libs"
 requires-python = ">= 3.11, <=3.12"
 version = "0.1.0"
 dependencies = [
-    "typedb-driver>=3.8.0",
-    "fastapi>=0.128.0",
-    "pyzotero>=1.7.6",
-    "pydantic>=2.12.5",
-    "python-dotenv>=1.2.1",
-    "anyio>=4.12.1",
-    "spacy>=3.8.11",
-    "loguru>=0.7.3",
-    "torch",
-    "torchvision",
-    "testcontainers",
-    "qdrant-client",
-    "docling>=2.73.0,<3",
-    "mkdocs-material",
-    "mkdocs",
-    "mkdocstrings[python]>=0.18",
-    "pypdf>=4.0.0"
-  , "openai>=2.29.0,<3", "transformers>=4.57.6,<5"]
+  "typedb-driver>=3.8.0",
+  "fastapi>=0.128.0",
+  "pyzotero>=1.7.6",
+  "pydantic>=2.12.5",
+  "python-dotenv>=1.2.1",
+  "anyio>=4.12.1",
+  "spacy>=3.8.11",
+  "loguru>=0.7.3",
+  "torch",
+  "torchvision",
+  "testcontainers",
+  "qdrant-client",
+  "docling>=2.73.0,<3",
+  "mkdocs-material",
+  "mkdocs",
+  "mkdocstrings[python]>=0.18",
+  "pypdf>=4.0.0",
+  "openai>=2.29.0,<3",
+  "transformers>=4.57.6,<5",
+]
 
 [dependency-groups]
 dev = [
@@ -47,10 +49,12 @@ requires = ["hatchling"]
 packages = ["src/database_builder_libs"]
 
 [tool.pixi.workspace]
-channels = ["https://fast.prefix.dev/conda-forge",
+channels = [
+  "https://fast.prefix.dev/conda-forge",
   "https://repo.prefix.dev/pypdfium2-public",
-  "https://repo.prefix.dev/pytorch"]
-platforms = ["linux-64","osx-arm64", "win-64"]
+  "https://repo.prefix.dev/pytorch",
+]
+platforms = ["linux-64", "osx-arm64", "win-64"]
 
 [tool.pixi.dependencies]
 # Dotenv is for reading .env files
@@ -65,13 +69,11 @@ database_builder_libs = { path = ".", editable = true }
 pythonpath = ["src"]
 addopts = ["--import-mode=importlib"]
 testpaths = ["tests/database_builder_libs"]
-filterwarnings = [
-    "ignore:datetime.datetime.utcnow.*:DeprecationWarning"
-]
+filterwarnings = ["ignore:datetime.datetime.utcnow.*:DeprecationWarning"]
 
 [tool.ty.src]
-include=["src"]
-exclude=["tests/", ".pixi/"]
+include = ["src"]
+exclude = ["tests/", ".pixi/"]
 
 [tool.ruff.lint]
 select = ["B024", "E4", "E7", "E9", "E301", "F", "C4", "PT"]
